editing file-bg.fish

+++ file-bg.fish
@@ -6,10 +6,17 @@ function file-bg
     set category (echo $target | cut -d "/" -f1)
     set name (echo $target | cut -d "/" -f2)
     set dir (get-folder-for-backgrounds $category)
-    # if not exists $dir
-    #   return 1
-    # end
+
+    if not test -d $dir
+        echo $dir does not exist
+        return 1
+    end
+
     set location $dir/$name.$ext
+    if test -e $location
+        echo target already exists
+        return 1
+    end
     mv $file $location
     wp $location
 end
